ref(runtime): Simplify retained broker execution
Remove unselected native Turn execution branches, localize session manager late ownership, and move Fleet adapter policy into the program owner. Reduce future snapshot dependencies to code needed inside broker Sandboxes.
