+++ scripts/daytona_snapshot.py
@@ -28,6 +28,7 @@ from fleet_rlm.daytona.provisioning import (
     DaytonaSandboxSpec,
     build_snapshot_image,
     environment_manifest,
+    snapshot_dependency_import_names,
 )
 
 
@@ -164,7 +165,7 @@ async def verify_runtime(client: Any, spec: DaytonaSandboxSpec) -> None:
             expected = json.dumps(manifest.image_identity(), sort_keys=True, separators=(",", ":"))
             compatible_profiles = tuple(profile.value for profile in manifest.compatible_profiles)
             code = (
-                "import getpass, hashlib, importlib.util, json, pathlib, shutil, sys, sysconfig\n"
+                "import getpass, hashlib, importlib, importlib.metadata, json, pathlib, shutil, sys, sysconfig\n"
                 "manifest_path = pathlib.Path('/opt/fleet/runtime-manifest.json')\n"
                 "manifest = json.loads(manifest_path.read_text())\n"
                 f"expected = json.loads({expected!r})\n"
@@ -179,9 +180,11 @@ async def verify_runtime(client: Any, spec: DaytonaSandboxSpec) -> None:
                 f"{manifest.digest!r}\n"
                 "assert sys.version_info[:3] == (3, 13, 13)\n"
                 "assert pathlib.Path(sys.executable).is_file()\n"
-                "dspy_spec = importlib.util.find_spec('dspy')\n"
-                "assert dspy_spec is not None and dspy_spec.origin\n"
-                "assert pathlib.Path(dspy_spec.origin).resolve().is_relative_to("
+                f"dependencies = {snapshot_dependency_import_names(spec.profile)!r}\n"
+                "for distribution, module, version in dependencies:\n"
+                "    imported = importlib.import_module(module)\n"
+                "    assert importlib.metadata.version(distribution) == version\n"
+                "    assert pathlib.Path(imported.__file__).resolve().is_relative_to("
                 "pathlib.Path(sysconfig.get_paths()['purelib']).resolve())\n"
                 "assert getpass.getuser() == 'daytona'\n"
                 "assert pathlib.Path.cwd() == pathlib.Path('/home/daytona')\n"
