+++ src/fleet_rlm/daytona/snapshot-requirements.txt
@@ -1,4 +1,3 @@
-dspy==3.3.1
 mpmath==1.4.1
 numpy==2.5.1
 pandas==3.0.5
