+++ src/fleet_rlm/rlm/events.py
@@ -23,7 +23,8 @@ import dspy
 from fleet_rlm.json_types import JsonValue, validate_json_value
 from fleet_rlm.observability.diagnostics import trace_failure_category
 from fleet_rlm.observability.tracing import turn_phase_span
-from fleet_rlm.rlm.compat_3_3_1 import FleetJSONAdapter, _RLMTraceCallback, is_native_rlm
+from fleet_rlm.rlm.compat_3_3_1 import _RLMTraceCallback, is_native_rlm
+from fleet_rlm.rlm.program import FleetJSONAdapter
 from fleet_rlm.rlm.result import (
     ExecutionDetail,
     RLMConfigError,
