+++ tests/unit/backend/rlm/test_adapter_budget_integration.py
@@ -10,8 +10,8 @@ import pytest
 from dspy.utils.exceptions import LMServerError
 
 from fleet_rlm.rlm.budget import AdapterBudget, BudgetDimension, BudgetLimits, TurnBudget, TurnBudgetExhausted
-from fleet_rlm.rlm.compat_3_3_1 import FleetJSONAdapter, _RLMTraceCallback
-from fleet_rlm.rlm.program import RLMModelBundle
+from fleet_rlm.rlm.compat_3_3_1 import _RLMTraceCallback
+from fleet_rlm.rlm.program import FleetJSONAdapter, RLMModelBundle
 from tests.unit.backend.rlm.test_fleet_json_adapter import _IterationActionSignature, _ScriptedLM
 
 GOOD = '{"reasoning":"done","code":"SUBMIT(answer=1)"}'
