+++ tests/unit/scripts/test_daytona_snapshot.py
@@ -8,7 +8,7 @@ import pytest
 from daytona import Image
 from daytona._async.snapshot import AsyncSnapshotService
 
-from fleet_rlm.daytona.provisioning import DaytonaSandboxSpec
+from fleet_rlm.daytona.provisioning import DaytonaSandboxSpec, snapshot_dependency_import_names
 from scripts import daytona_snapshot
 
 
@@ -34,6 +34,61 @@ def test_help_needs_no_credentials(capsys: pytest.CaptureFixture[str]) -> None:
     with pytest.raises(SystemExit, match="0"):
         daytona_snapshot.main(["--help"])
     assert "immutable Fleet Daytona Snapshot" in capsys.readouterr().out
+
+
+@pytest.mark.asyncio
+@pytest.mark.parametrize("profile", ["session", "semantic-child"])
+@pytest.mark.parametrize("succeeds", [True, False])
+async def test_runtime_probe_checks_profile_dependencies_and_cleans_up(
+    monkeypatch: pytest.MonkeyPatch, profile: str, succeeds: bool
+) -> None:
+    spec = daytona_snapshot._spec("fleet-probe-v1", profile)
+    operations: list[str] = []
+    context = object()
+
+    async def create_context():
+        return context
+
+    async def run_code(code, *, context):
+        assert context is not None
+        compile(code, "snapshot-probe", "exec")
+        assert "dspy" not in code
+        assert f"dependencies = {snapshot_dependency_import_names(spec.profile)!r}" in code
+        assert "importlib.import_module(module)" in code
+        assert "importlib.metadata.version(distribution) == version" in code
+        return SimpleNamespace(stdout="fleet-snapshot-runtime-ok" if succeeds else "", error=None)
+
+    async def delete_context(received):
+        assert received is context
+        operations.append("delete-context")
+
+    sandbox = SimpleNamespace(
+        id="probe",
+        code_interpreter=SimpleNamespace(
+            create_context=create_context, run_code=run_code, delete_context=delete_context
+        ),
+    )
+
+    async def create(**_kwargs):
+        return sandbox
+
+    async def delete(received):
+        assert received is sandbox
+        operations.append("delete-sandbox")
+
+    async def get(sandbox_id):
+        assert sandbox_id == "probe"
+        operations.append("confirm-absence")
+        return None
+
+    platform = SimpleNamespace(create=create, delete=delete, get=get)
+    monkeypatch.setattr(daytona_snapshot, "LiveDaytonaPlatform", lambda *_args: platform)
+    if succeeds:
+        await daytona_snapshot.verify_runtime(object(), spec)
+    else:
+        with pytest.raises(RuntimeError, match="snapshot runtime probe failed"):
+            await daytona_snapshot.verify_runtime(object(), spec)
+    assert operations == ["delete-context", "delete-sandbox", "confirm-absence"]
 
 
 @pytest.mark.asyncio
